api/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
psycopg[binary,pool]==3.2.12
python-dotenv==1.2.1
argon2-cffi==25.1.0
flask-cors==6.0.1
requests==2.32.5